pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "black>=25.1.0",
    "fastapi>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
    "websockets>=15.0.1",
]

[tool.pyright]
typeCheckingMode = "off"
venv = ".venv"
venvPath = "."

[tool.ruff]
line-length = 88

[tool.ruff.format]
quote-style = "single"
